+++ sandbox/src/main/scala/example/WebSockets.scala
@@ -10,9 +10,10 @@ import tyrian.Task.Observable
 import tyrian.Task.Observer
 import tyrian.Task.Cancelable
 
-trait WebSockets:
-
-  private val sockets: mutable.HashMap[WebSocketId, LiveSocket] = mutable.HashMap()
+// TODO: A reconnect back off policy
+// TODO: A number of connection attempts before abort
+// TODO: A way to trigger a manual reconnect
+final class WebSockets(sockets: mutable.HashMap[WebSocketId, LiveSocket]):
 
   // send cmd
   def send[Msg](id: WebSocketId, message: String, onError: => Msg): Cmd[Msg] =
@@ -82,4 +83,34 @@ trait WebSockets:
         )
     }
 
-  final case class LiveSocket(socket: dom.WebSocket, subs: Sub[WebSocketEvent])
+object WebSockets:
+  def apply(): WebSockets =
+    new WebSockets(mutable.HashMap())
+
+final case class LiveSocket(socket: dom.WebSocket, subs: Sub[WebSocketEvent])
+
+enum WebSocketEvent derives CanEqual:
+  case Open(webSocketId: WebSocketId)                     extends WebSocketEvent
+  case Receive(webSocketId: WebSocketId, message: String) extends WebSocketEvent
+  case Error(webSocketId: WebSocketId, error: String)     extends WebSocketEvent
+  case Close(webSocketId: WebSocketId)                    extends WebSocketEvent
+
+opaque type WebSocketId = String
+object WebSocketId:
+  inline def apply(id: String): WebSocketId          = id
+  extension (wsid: WebSocketId) def toString: String = wsid
+
+final case class WebSocketConfig(id: WebSocketId, address: String) derives CanEqual
+
+enum WebSocketReadyState derives CanEqual:
+  case CONNECTING, OPEN, CLOSING, CLOSED
+
+object WebSocketReadyState:
+  def fromInt(i: Int): WebSocketReadyState =
+    i match {
+      case 0 => CONNECTING
+      case 1 => OPEN
+      case 2 => CLOSING
+      case 3 => CLOSED
+      case _ => CLOSED
+    }
